+++ NEWS.md
@@ -4,6 +4,11 @@
 
 -   Shortened global model list to just xgboost
 -   Faster xgboost model training for larger datasets 
+-   Faster feature selection for global model training
+
+## Bug Fixes
+
+-   Error in formatting of training data for global models
 
 # finnts 0.5.0